Migrate legacy API from rtm.start to rtm.connect
This makes the client mostly compatible with the September 2022 update
